feat: refactor tests package form feelExpression to expression
