+++ zeebe/protocol-impl/src/test/resources/protocol/records/golden/ExpressionRecord.golden
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.protocol.impl.record.value.feelexpression;
+package io.camunda.zeebe.protocol.impl.record.value.expression;
 
 import io.camunda.zeebe.msgpack.property.ArrayProperty;
 import io.camunda.zeebe.msgpack.property.BinaryProperty;
